requirements.txt
cairocffi==0.9.0
cffi==1.11.5
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
opencv-python==3.4.3.18
pgi==0.0.11.2
Pillow==5.2.0
pycairo==1.17.1
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
ruamel.yaml==0.15.66
six==1.11.0
vext==0.7.0
vext.gi==0.7.0
